[CPDLP-4009] Renamed Contract component to Contract::ECF
