--- app/components/finance/statements/contract_band_row.html.erb
@@ -1,6 +1,0 @@
-<tr class="govuk-table__row">
-  <td class="govuk-table__cell"><%= t("finance.ecf.bands.#{index}") %></td>
-  <td class="govuk-table__cell govuk-table__cell--numeric"><%= min || t("finance.contract.bands.floor") %></td>
-  <td class="govuk-table__cell govuk-table__cell--numeric"><%= max || t("finance.contract.bands.ceiling") %></td>
-  <td class="govuk-table__cell govuk-table__cell--numeric"><%= number_to_pounds per_participant %></td>
-</tr>
