+++ app/components/finance/statements/contracts/ecf.html.erb
@@ -63,17 +63,22 @@
 
     <table class="govuk-table">
       <thead class="govuk-table__head">
-      <tr class="govuk-table__row">
-        <th class="govuk-table__header"><%=t "finance.contract.band" %></th>
-        <th class="govuk-table__header govuk-table__cell--numeric"><%=t "finance.contract.bands.min" %></th>
-        <th class="govuk-table__header govuk-table__cell--numeric"><%=t "finance.contract.bands.max" %></th>
-        <th class="govuk-table__header govuk-table__cell--numeric"><%=t "finance.per_participant"%></th>
-      </tr>
+        <tr class="govuk-table__row">
+          <th class="govuk-table__header"><%=t "finance.contract.band" %></th>
+          <th class="govuk-table__header govuk-table__cell--numeric"><%=t "finance.contract.bands.min" %></th>
+          <th class="govuk-table__header govuk-table__cell--numeric"><%=t "finance.contract.bands.max" %></th>
+          <th class="govuk-table__header govuk-table__cell--numeric"><%=t "finance.per_participant"%></th>
+        </tr>
       </thead>
       <tbody class="govuk-table__body">
-
-      <%= render Finance::Statements::ContractBandRow.with_collection(bands) %>
-
+        <% bands.each_with_index.map do |band, index| %>
+          <tr class="govuk-table__row">
+            <td class="govuk-table__cell"><%= t("finance.ecf.bands.#{index}") %></td>
+            <td class="govuk-table__cell govuk-table__cell--numeric"><%= band.min || t("finance.contract.bands.floor") %></td>
+            <td class="govuk-table__cell govuk-table__cell--numeric"><%= band.max || t("finance.contract.bands.ceiling") %></td>
+            <td class="govuk-table__cell govuk-table__cell--numeric"><%= number_to_pounds band.per_participant %></td>
+          </tr>
+        <% end %>
       </tbody>
     </table>
   </div>
